moving Heartbeat and adding a better way to redirect streams
